Revert "refactor: move overrides into per-doctype folders"
This reverts commit b3b717d183c2a3432c1f82888c0b2845928cecf8.
